requirements.txt
flask
pinecone-client
openai
python-dotenv